Add tile.terrain value and include water tiles
Water tiles have a terrain value < 0 and aren't drawn by the client.
Disallowing movement to water tiles is only done client-side and is yet
another thing that needs to be validated in LoadCommand().

The tile.connected values for water tiles need to be kept to allow for
bridges/boats etc...